requirements.txt
# Core AI/ML
langchain==0.1.0
langchain-community==0.0.20
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# Vector Stores
chromadb==0.4.18
pinecone-client==2.2.4  # Optional: voor productie
faiss-cpu==1.7.4

# Document Processing
pypdf==3.17.4
tiktoken==0.5.2
sentence-transformers==2.2.2

# Data Processing (FIXED NUMPY VERSION)
pandas==2.0.3
numpy==1.24.3

# Web Framework
streamlit==1.29.0
fastapi==0.105.0
uvicorn==0.25.0

# Visualization
plotly==5.18.0
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6

# Text Processing
spacy==3.7.2
nltk==3.8.1

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Production
gunicorn==21.2.0
redis==5.0.1